refect: Controller readline처리 View로 변경
